projects/face_recognition/requirements.txt
facenet-pytorch
opencv-python
python-dotenv
SQLAlchemy
psycopg2-binary